added serve to watch
